#1006 Compile warning free with stricter gcc/clang warnings
* Make code compile with stricter gcc/clang warning settings

* Fix formatting

* Fix GCC warnings

+++ test/custom_builds/c/pipeline/src/main.c
@@ -1,7 +1,7 @@
 #include <pipeline.h>
 
 typedef struct {
-    float x, y;
+    int32_t x, y;
 } Position, Velocity;
 
 void Move(ecs_iter_t *it) {
