+++ test/custom_builds/c/pipeline_w_system/src/main.c
@@ -1,7 +1,7 @@
 #include <pipeline_w_system.h>
 
 typedef struct {
-    float x, y;
+    int32_t x, y;
 } Position, Velocity;
 
 void Move(ecs_iter_t *it) {
