+++ test/custom_builds/cpp/pipeline/src/main.cpp
@@ -2,11 +2,11 @@
 #include <iostream>
 
 struct Position { 
-    double x, y;
+    int32_t x, y;
 };
 
 struct Velocity { 
-    double x, y;
+    int32_t x, y;
 };
 
 int main(int, char *[]) {
